improvements in 'index.hmtl'
