+++ bld/qtcreator/container/tst/tst.pro
@@ -9,9 +9,10 @@ PRJ_DIR=$$BASE_DIR/tnct/container/tst
 
 SOURCES = $$PRJ_DIR/main.cpp
 
-HEADERS = $$PRJ_DIR/matrix_test.h \
+HEADERS = \
           $$PRJ_DIR/circular_queue_test.h \
-          $$PRJ_DIR/cpt_test.h
+          $$PRJ_DIR/cpt_test.h \
+          $$PRJ_DIR/matrix_test.h
 
 DISTFILES += \
     $$PRJ_DIR/circular_queue_tests_000.ini
