+++ src/tnct/crosswords/doc/jogos/plantas-fr.cross
@@ -1,0 +1,15 @@
+Pommier|Arbre qui produit des pommes
+Bouleau|Arbre typique des régions froides
+Chêne|Arbre symbole de force
+Sapin|Arbre utilisé comme arbre de Noël
+Pissenlit|Plante à fleurs jaunes et graines volantes
+Châtaignier|Arbre qui produit des fruits comestibles et des bogues épineuses
+Vigne|Plante qui produit des raisins
+Tournesol|Fleur qui suit le mouvement du soleil
+Lavande|Plante aromatique aux fleurs lilas
+Romarin|Herbe aromatique utilisée en cuisine
+Sauge|Herbe médicinale et aromatique
+Basilic|Herbe utilisée dans les plats méditerranéens
+Persil|Herbe utilisée comme condiment
+Menthe|Herbe aromatique rafraîchissante
+Trèfle|Plante herbacée souvent associée à la chance
