+++ src/tnct/crosswords/tst/assembler_test.h
@@ -24,6 +24,16 @@ using namespace std::chrono_literals;
 namespace tnct::crosswords::tst
 {
 
+std::uint64_t number_of_permutations(const dat::entries &p_entries)
+{
+  if (p_entries.get_num_entries() > 20)
+  {
+    return math::factorial<std::uint64_t>(20).value();
+  }
+
+  return math::factorial<std::uint64_t>(p_entries.get_num_entries()).value();
+}
+
 struct assembler_test_000
 {
   static std::string desc()
@@ -52,7 +62,8 @@ struct assembler_test_000
     assembler _solver(_logger, _dispatcher);
 
     auto _start{std::chrono::high_resolution_clock::now()};
-    auto _grid{_solver.start(_entries, index{11}, index{11}, 1s)};
+    auto _grid{_solver.start(_entries, index{11}, index{11}, 1s,
+                             number_of_permutations(_entries))};
     auto _end{std::chrono::high_resolution_clock::now()};
     std::chrono::duration<double> diff = _end - _start;
     TNCT_LOG_TST(_logger, format::fmt("time: ", diff.count()));
@@ -96,7 +107,8 @@ struct assembler_test_001
 
     auto _start{std::chrono::high_resolution_clock::now()};
     auto _grid{_solver.start(_entries, crosswords::dat::index{11},
-                             crosswords::dat::index{11}, 1s)};
+                             crosswords::dat::index{11}, 1s,
+                             number_of_permutations(_entries))};
     auto _end{std::chrono::high_resolution_clock::now()};
     std::chrono::duration<double> diff = _end - _start;
     TNCT_LOG_TST(_logger, format::fmt("time: ", diff.count()));
@@ -148,7 +160,8 @@ struct assembler_test_002
 
     auto _start{std::chrono::high_resolution_clock::now()};
     auto _grid{_solver.start(_entries, crosswords::dat::index{11},
-                             crosswords::dat::index{11}, 2s)};
+                             crosswords::dat::index{11}, 2s,
+                             number_of_permutations(_entries))};
     auto _end{std::chrono::high_resolution_clock::now()};
     std::chrono::duration<double> diff = _end - _start;
     TNCT_LOG_TST(_logger, format::fmt("time: ", diff.count()));
