--- src/tnct/crosswords/uix/client/html/index.html.00
@@ -1,173 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<meta charset="UTF-8">
-<title>Crossword Client</title>
-<style>
-body { font-family: Arial, sans-serif; margin: 20px; }
-label { display: inline-block; width: 200px; }
-input[type="text"] { width: 80px; margin-bottom: 5px; }
-button { margin: 5px; }
-canvas { border: 1px solid black; margin-top: 20px; }
-table { border-collapse: collapse; width: 100%; }
-th, td { border: 1px solid black; padding: 4px; }
-th.idCol, td.idCol { text-align: right; }
-#clues { display: flex; gap: 20px; margin-top: 20px; }
-#clues > div { flex: 1; }
-</style>
-</head>
-<body>
-<h1>Crossword Client</h1>
-
-<button id="loadButton">Load</button>
-<input type="file" id="fileInput" accept=".txt" style="display:none" />
-
-<div>
-  <label>Number of Rows:</label><input type="text" id="rowsInput"><br>
-  <label>Number of Cols:</label><input type="text" id="colsInput"><br>
-  <label>Maximum number of Rows:</label><input type="text" id="maxRowsInput"><br>
-  <label>Interval (secs):</label><input type="text" id="intervalInput"><br>
-</div>
-
-<button id="submitButton">Submit</button>
-<button id="printButton">Print</button>
-
-<h2>Incoming Messages</h2>
-<pre id="messages"></pre>
-
-<div id="printArea">
-  <canvas id="gridCanvas" width="600" height="600"></canvas>
-  <div id="clues">
-    <div>
-      <h3>Horizontals</h3>
-      <table id="horizontalClues">
-        <tr><th class="idCol">ID</th><th>Explanation</th></tr>
-      </table>
-    </div>
-    <div>
-      <h3>Verticals</h3>
-      <table id="verticalClues">
-        <tr><th class="idCol">ID</th><th>Explanation</th></tr>
-      </table>
-    </div>
-  </div>
-</div>
-
-<script>
-let ws;
-let entries = [];
-let rows=0, cols=0, max_rows=0, interval=0;
-
-// WebSocket connect on page load
-function connectWebSocket() {
-  ws = new WebSocket("ws://localhost:9002");
-  ws.onopen = () => console.log("WebSocket connected");
-  ws.onerror = e => console.error("WebSocket error", e);
-  ws.onmessage = event => {
-    try {
-      const msg = JSON.parse(event.data);
-      if(msg.response==="solved") {
-        document.getElementById("messages").textContent = JSON.stringify(msg, null,2);
-        renderGrid(msg);
-        renderClues(msg.layouts);
-      }
-    } catch(err) { console.error("Invalid message", err); }
-  };
-}
-window.addEventListener("load", connectWebSocket);
-
-// Load button
-document.getElementById("loadButton").onclick = () => document.getElementById("fileInput").click();
-document.getElementById("fileInput").addEventListener("change", e => {
-  const file = e.target.files[0];
-  if(!file) return;
-  const reader = new FileReader();
-  reader.onload = ev => {
-    entries = ev.target.result
-      .split("\n").map(l=>l.trim()).filter(l=>l.includes("|"))
-      .map(l=>{const [word, explanation]=l.split("|"); return {word:word.trim(), explanation:explanation.trim()};});
-    console.log("Entries loaded:", entries);
-    alert("Entries loaded! Click Submit to send.");
-  };
-  reader.readAsText(file);
-});
-
-// Inputs
-document.getElementById("rowsInput").oninput = e => rows = parseInt(e.target.value)||0;
-document.getElementById("colsInput").oninput = e => cols = parseInt(e.target.value)||0;
-document.getElementById("maxRowsInput").oninput = e => max_rows = parseInt(e.target.value)||0;
-document.getElementById("intervalInput").oninput = e => interval = parseInt(e.target.value)||0;
-
-// Submit
-document.getElementById("submitButton").onclick = () => {
-  if(!ws || ws.readyState!==WebSocket.OPEN){ alert("WebSocket not connected"); return; }
-  ws.send(JSON.stringify({request:"create",rows,cols,max_rows,interval,entries}));
-};
-
-// Print only grid + tables
-document.getElementById("printButton").onclick = () => {
-  const canvas = document.getElementById("gridCanvas");
-  const imgData = canvas.toDataURL("image/png");
-  const horizTable = document.getElementById("horizontalClues").outerHTML;
-  const vertTable = document.getElementById("verticalClues").outerHTML;
-  const printWindow = window.open("", "_blank");
-  printWindow.document.write("<html><head><title>Print</title></head><body>");
-  printWindow.document.write(`<img src="${imgData}" style="max-width:100%"><br><br>`);
-  printWindow.document.write("<div style='display:flex; gap:20px;'>");
-  printWindow.document.write("<div style='flex:1'>" + horizTable + "</div>");
-  printWindow.document.write("<div style='flex:1'>" + vertTable + "</div>");
-  printWindow.document.write("</div></body></html>");
-  printWindow.document.close();
-  printWindow.onload = () => { printWindow.print(); printWindow.close(); };
-};
-
-// Render grid
-function renderGrid(msg){
-  const canvas = document.getElementById("gridCanvas");
-  const ctx = canvas.getContext("2d");
-  ctx.clearRect(0,0,canvas.width,canvas.height);
-  const nRows = parseInt(msg.rows.replace(/"/g,""));
-  const nCols = parseInt(msg.cols.replace(/"/g,""));
-  const cellSize = Math.min(canvas.width/nCols, canvas.height/nRows);
-
-  ctx.strokeStyle = "black";
-  ctx.font = `${cellSize/3}px Arial`;
-  ctx.textBaseline = "top";
-
-  // Draw grid
-  for(let r=0;r<nRows;r++){
-    for(let c=0;c<nCols;c++){
-      ctx.strokeRect(c*cellSize,r*cellSize,cellSize,cellSize);
-    }
-  }
-
-  // Place IDs
-  msg.layouts.forEach(layout=>{
-    const x = parseInt(layout.col)*cellSize;
-    const y = parseInt(layout.row)*cellSize;
-    if(layout.orientation==="V"){
-      ctx.textAlign="left";
-      ctx.fillText(layout.id||"", x+2, y+2);
-    } else if(layout.orientation==="H"){
-      ctx.textAlign="right";
-      ctx.fillText(layout.id||"", x+cellSize-2, y+2);
-    }
-  });
-}
-
-// Render clues
-function renderClues(layouts){
-  const horizTable = document.getElementById("horizontalClues");
-  const vertTable = document.getElementById("verticalClues");
-  horizTable.innerHTML="<tr><th class='idCol'>ID</th><th>Explanation</th></tr>";
-  vertTable.innerHTML="<tr><th class='idCol'>ID</th><th>Explanation</th></tr>";
-  layouts.forEach(l=>{
-    const row = `<tr><td class='idCol'>${l.id||""}</td><td>${l.explanation}</td></tr>`;
-    if(l.orientation==="H") horizTable.innerHTML += row;
-    else vertTable.innerHTML += row;
-  });
-}
-</script>
-</body>
-</html>
-
